Separated functionality of processing json file and actually displaying the graph
